refactor(app): move infra wiring into dedicated modules
Keep AppModule focused on composition by extracting logging,
queue bootstrap, and Bull Board admin setup into
src/modules/infra modules.

Made-with: Cursor
